move the files to the right place Co-authored-by: M Daley <mdaley-madetech@users.noreply.github.com>
