+++ server/routes/establishments/wdfClaims/grantLetter.js
@@ -1,16 +1,8 @@
 const router = require('express').Router();
 const { hasPermission } = require('../../../utils/security/hasPermission');
+const { generateDevelopmentFundGrantLetter } = require('./generateDevelopmentFundGrantLetter');
 
-const updateBUDataChanges = async (res) => {
-  try {
-    res.status(200).send();
-  } catch (error) {
-    console.log(error);
-    res.status(500).send();
-  }
-};
-
-router.route('/').post(hasPermission('canManageWdfClaims'), updateBUDataChanges);
+router.route('/').post(generateDevelopmentFundGrantLetter);
 
 module.exports = router;
-module.exports.updateBUDataChanges = updateBUDataChanges;
+module.exports.generateDevelopmentFundGrantLetter = generateDevelopmentFundGrantLetter;
