+++ server/routes/wdf/index.js
@@ -3,9 +3,7 @@ const express = require('express');
 const router = express.Router();
 
 const DisbursementFile = require('./disbursementFile');
-const DevelopmentFund = require('./developmentFundGrants');
 
 router.use('/disbursementFile', DisbursementFile);
-router.use('/developmentFund', DevelopmentFund);
 
 module.exports = router;
